fix: login page title, quantity type, material->unit relation, namespace and frontend issues

+++ Modules/SCM/Database/Migrations/2023_02_05_042408_create_scm_requisition_details_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->integer('material_id')->nullable();
             $table->string('description')->nullable();
             $table->string('item_code')->nullable();
-            $table->integer('quantity')->nullable();
+            $table->double('quantity', 8, 2)->nullable();
             $table->integer('brand_id')->nullable();
             $table->string('model')->nullable();
             $table->string('purpose')->nullable();
